Checks similar names of output PDF files
Added a condition to check similar file names in the text file in order to prevent over-writing of the files of similar names. If there are more than one files of same name listed in text file, page range is added in the end of the file name to distinguish such files.